Implemented TODOs
