+++ StackExRedis/NHibernate.Caches.StackExRedis/DefaultCacheRegionStrategyFactory.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using System.Collections.Generic;
+using NHibernate.Cache;
 using StackExchange.Redis;
 
 namespace NHibernate.Caches.StackExRedis
@@ -8,11 +9,21 @@ namespace NHibernate.Caches.StackExRedis
 	public class DefaultCacheRegionStrategyFactory : ICacheRegionStrategyFactory
 	{
 		/// <inheritdoc />
-		public AbstractRegionStrategy Create(IConnectionMultiplexer connectionMultiplexer,
+		public virtual AbstractRegionStrategy Create(IConnectionMultiplexer connectionMultiplexer,
 			RedisCacheRegionConfiguration configuration, IDictionary<string, string> properties)
 		{
-			return (AbstractRegionStrategy) Activator.CreateInstance(configuration.RegionStrategy, connectionMultiplexer,
-				configuration, properties);
+			if (configuration.RegionStrategy == typeof(DefaultRegionStrategy))
+			{
+				return new DefaultRegionStrategy(connectionMultiplexer, configuration, properties);
+			}
+			if (configuration.RegionStrategy == typeof(FastRegionStrategy))
+			{
+				return new FastRegionStrategy(connectionMultiplexer, configuration, properties);
+			}
+
+			throw new CacheException(
+				$"{configuration.RegionStrategy} is not supported by {GetType()}, register " +
+				$"a custom {typeof(ICacheRegionStrategyFactory)} or use a supported one.");
 		}
 	}
 }
