+++ StackExRedis/NHibernate.Caches.StackExRedis/NHibernate.Caches.StackExRedis.csproj
@@ -32,4 +32,7 @@
       <PackagePath>./NHibernate.Caches.license.txt</PackagePath>
     </Content>
   </ItemGroup>
+  <ItemGroup>
+    <ProjectReference Include="..\..\NHibernate.Caches.Common\NHibernate.Caches.Common.csproj" />
+  </ItemGroup>
 </Project>
